+++ Back End/Curriculum Calculator/Student.h
@@ -19,14 +19,12 @@ class Student
     void addMajor(Major);
     void removeMajor(int);
     std::vector<long> getTaken();
-    void addTaken(int);
-    void removeTaken(int);
+    void addTaken(long);
+    void removeTaken(long);
     void calculateRequired();
     void calculateRemaining();
     void calculateRecommended();
     std::vector<std::vector<long>> getRequired();
     std::vector<std::vector<long>> getRemaining();
     std::vector<std::vector<long>> getRecommended();
-    
-
 };
